Fix screenshot offset as well as screenshot aspect ratio, also use PNG instead of BMP thanks to issue 1186. Fix for opening.bnr containing bad chars where the loop would stop when encountering one (fixes Rogue Leader description, which was using 0x0d and 0x0a)
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@3899 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Core/Src/IPC_HLE/WII_IPC_HLE_Device_net.cpp
@@ -330,7 +330,7 @@ u32 CWII_IPC_HLE_Device_net_ip_top::ExecuteCommand(u32 _Command, u32 _BufferIn, 
 		struct sockaddr_in address;
 		int Return = accept(S, (struct sockaddr *)&address, &addrlen);
 		GC_sockaddr_in *addr = (GC_sockaddr_in*)Memory::GetPointer(BufferOut);
-		addr->sin_family = address.sin_family;
+		addr->sin_family = (u8)address.sin_family;
 		addr->sin_addr.s_addr_ = address.sin_addr.s_addr;
 		addr->sin_port = address.sin_port;
 		socklen_t *Len = (socklen_t *)Memory::GetPointer(BufferOut + 0x04);
